Initial modifications for Rev5 hardware
